fix: frames no longer drop rest of message (#3)
According to what I could find in the docs,
only bitstrings are allowed to take a -1
length and then match the rest of the bitstring.

So the string was always just empty and no "rest"
was ever found. This caused messages to be dropped
if they arrived in the same packet (for example,
only deserialize the Ping message. Which is bad).